requirements.txt
matplotlib
numpy
torch
Pillow
torchvision
